Handle add & edit

+++ src/components/Products.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
-
+import Swal from 'sweetalert2'
 export default function Products() {
   const apiLink = 'http://localhost:3000/products'
   const [products, setProducts] = useState([])
@@ -9,17 +9,24 @@ export default function Products() {
   useEffect(() => {
     getData()
   }, [])
- function getData(){
-  fetch(apiLink).then((respons) => respons.json()).then((data) => setProducts(data))
-
- }
+  function getData() {
+    fetch(apiLink).then((respons) => respons.json()).then((data) => setProducts(data))
+  }
 
   // to delete
-  function deleteProduct(id) {
-    fetch(`${apiLink}/${id}`,{method: 'delete'}).then((respons) => respons.json()).then((data) => getData() )
+  function deleteProduct(id, title) {
+    Swal.fire({
+      title: `Are You sure to delete ${title}?`,
+      icon: "question",
+      showCancelButton: true,
+    }).then((data) => {
+      if (data.isConfirmed == true) {
+        fetch(`${apiLink}/${id}`, { method: 'delete' }).then((respons) => respons.json()).then((data) => getData())
+      }
+    })
 
   }
-  
+
   return <>
     <div className='d-flex p-3 mb-4 w-100 justify-space-between align-items-center'>
       <h1 className='d-inline'>Products List</h1>
@@ -43,11 +50,12 @@ export default function Products() {
             <td>{e.title}</td>
             <td className='text-center'>{e.price}</td>
             <td className='text-center'>
-              <button className='btn btn-danger btn-sm' onClick={() => deleteProduct(`${e.id}`)}>Delete</button>
+              <button className='btn btn-danger btn-sm' onClick={() => deleteProduct(`${e.id}`, `${e.title}`)}>Delete</button>
               <Link to={`${e.id}`}>
                 <button className='btn btn-info btn-sm mx-3'>View</button>
               </Link>
-              <button className='btn btn-warning btn-sm'>Edit</button>
+              <Link to={`../edit/${e.id}`}>
+              <button className='btn btn-warning btn-sm'>Edit</button> </Link>
             </td>
           </tr>
         })}
